refactor(frontend/pages): rectify folder structure
