+++ frontend/app/(protected)/(tabs)/profile.tsx
@@ -50,7 +50,7 @@ export default function ProfileScreen() {
   return (
     <YStack f={1} ai="center" gap="$8" px="$10" pt="$5" bg="$background">
       <H2>Profile</H2>
-      <Link href="/Pages/services">
+      <Link href="/services">
         <Text>services</Text>
       </Link>
       <Link href="#" onPress={disconnect}>
