+++ frontend/app/(protected)/Create/actions.tsx
@@ -1,7 +1,7 @@
 import { Button, ScrollView, Stack, YStack, Text, View, XStack, Square, H2 } from 'tamagui'
-import { useNavigationData } from '../../../context/navigationContext'
+import { useNavigationData } from '../../context/navigationContext'
 import { Link } from 'expo-router'
-import { useApplet } from '../../../context/appletContext';
+import { useApplet } from '../../context/appletContext';
 import React from 'react';
 
 
@@ -22,7 +22,7 @@ export default function ServicesScreen() {
       <YStack paddingVertical="$4" width="100%" alignItems='center' gap="$2" >
         <H2>{navigationData.currentService}</H2>
         {actions.flatMap((action, i) => [
-            <Link key={`button-${action.id}`} href="/Pages/Create/form" asChild >
+            <Link key={`button-${action.id}`} href="/Create/form" asChild >
                 <Button
                     onPress={() => {setApplet(
                       {
